Prepare for package creation, move source code to under 'gliner' dir
