+++ gliner/__init__.py
@@ -1,0 +1,3 @@
+from .model import GLiNER
+
+__all__ = ["GLiNER"]
